Modify records in language table [#27 state:fixed responsible:none]
Language table had records about preinstalled languages it and en. For the
client it is needed to have a full identification of the language and
localization so instead it is now: en-US and it-IT.
No structure is changed!!!
From now on, it is necessary to define the language code as the
combination above.