Add support for map search configuration and layer visibility [#89 state:fixed responsible:none][#88 state:open responsible:none]
Ticket #89: Added table system.map_search_option to configure the search
options of the Map Search control.
The table is populated with 3 standard search options.
In the table query are added also 3 queries for each of these search
options.
Ticket #88: Added an extra column to the table system.config_map_layer
called visible_on_start which by default is true. This column defines if a
layer must be visible in start.