Add functions to check validity of ba_unit, cadastre_object names[#77]
It is added administrative.ba_unit_name_is_valid(name_firstpart : varchar,
name_lastpart : varchar). This function returns false if the name of the
ba_unit
presented by two parameters does not fullfil the requirements.
This function can be used in business rules to check for validity.
It is added in the Administrative functions in the data model.

It is added cadastre.cadastre_object_name_is_valid(name_firstpart :
varchar, name_lastpart : varchar).
This function returns false if the name of the cadastre_object presented
by two parameters does not fullfil the requirements.
It can be used in business rules to check for validity.
It is added in the Cadastre functions in the data model.