Reduce feature count based in pixelresolution[#68]
It is added in queries that retrieve features for parcels and roads an
extra condition to reduce the number of features based in the resolution
that is asking them.
So if there is a small zoom (meter per pixel is big) features of type
polygon (in this case parcels and roads) that have an area less than
resolution X 150 are left out.
So like this, the number of features asked for a certain area is kept
within a reasonable number.
In the model:
1. Go to record editor of the system.query table
2. Replace the sql value for queries: SpatialResult.getParcels and
SpatialResult.getRoads